backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
structlog==24.4.0
httpx==0.27.2
yfinance==0.2.44
numpy==1.26.4
pandas==2.2.3
torch==2.2.2
requests==2.32.3
python-dotenv==1.0.1
ccxt==4.3.89
pydantic-settings==2.5.2
